improve(hapi): changed hapi server routes structure
