+++ hapi/package.json
@@ -17,12 +17,14 @@
   "author": "EOS Costa Rica",
   "license": "MIT",
   "dependencies": {
+    "@hapi/boom": "^10.0.0",
     "@hapi/hapi": "^20.1.5",
     "axios": "^0.26.1",
     "eosjs": "^22.1.0",
     "eosjs-api": "^7.0.4",
     "graphql": "^16.3.0",
     "graphql-request": "^4.1.0",
+    "joi": "^17.7.0",
     "massive": "^6.9.1",
     "moment": "^2.29.1",
     "node-fetch": "^2.6.7",
