+++ hapi/src/routes/healthz.route.js
@@ -1,0 +1,10 @@
+module.exports = {
+  method: 'GET',
+  path: '/healthz',
+  handler: () => {
+    return '<h2>EOS Rate HTTP API service</h2>'
+  },
+  options: {
+    auth: false
+  }
+}
