+++ hapi/src/routes/index.js
@@ -1,0 +1,4 @@
+const healthzRoute = require('./healthz.route')
+const ratebpRoute = require('./ratebp.route')
+
+module.exports = [healthzRoute, ratebpRoute]
