+++ hasura/metadata/actions.yaml
@@ -1,16 +1,22 @@
 actions:
-- name: rateProducer
-  definition:
-    kind: synchronous
-    handler: '{{HASURA_GRAPHQL_ACTION_BASE_URL}}/ratebp'
-  permissions:
-  - role: anonymous
+  - name: getRatesStats
+    definition:
+      kind: synchronous
+      handler: '{{HASURA_GRAPHQL_ACTION_BASE_URL}}/get-rates-stats'
+  - name: rateProducer
+    definition:
+      kind: synchronous
+      handler: '{{HASURA_GRAPHQL_ACTION_BASE_URL}}/ratebp'
+    permissions:
+      - role: anonymous
 custom_types:
   enums: []
   input_objects:
-  - name: RatingInput
+    - name: RatingInput
+    - name: RatesStatsInput
   objects:
-  - name: RatingOutput
-  - name: deleteUserRateOutput
-  - name: AddCommentOutput
+    - name: RatingOutput
+    - name: deleteUserRateOutput
+    - name: AddCommentOutput
+    - name: RatesStatsOutput
   scalars: []
