+++ hasura/metadata/api_limits.yaml
@@ -1,0 +1,1 @@
+{}
