+++ hasura/metadata/databases/databases.yaml
@@ -2,13 +2,13 @@
   kind: postgres
   configuration:
     connection_info:
+      use_prepared_statements: true
       database_url:
         from_env: HASURA_GRAPHQL_DATABASE_URL
       isolation_level: read-committed
       pool_settings:
         connection_lifetime: 600
+        retries: 1
         idle_timeout: 180
         max_connections: 50
-        retries: 1
-      use_prepared_statements: true
   tables: "!include default/tables/tables.yaml"
