+++ hasura/metadata/databases/default/tables/public_hyperion_state.yaml
@@ -1,3 +1,3 @@
 table:
-  name: hyperion_state
   schema: public
+  name: hyperion_state
