+++ hasura/metadata/databases/default/tables/public_total_ratings_stats.yaml
@@ -1,3 +1,3 @@
 table:
-  name: total_ratings_stats
   schema: public
+  name: total_ratings_stats
