+++ hasura/metadata/graphql_schema_introspection.yaml
@@ -1,0 +1,1 @@
+disabled_for_roles: []
