+++ hasura/metadata/network.yaml
@@ -1,0 +1,1 @@
+{}
